Extract Internal API Usage filter to separate class
